+++ lib/nerves_hub_link/extensions/logging/logger_handler.ex
@@ -1,0 +1,23 @@
+# SPDX-FileCopyrightText: 2025 Josh Kalderimis
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+defmodule NervesHubLink.Extensions.Logging.LoggerHandler do
+  @moduledoc """
+  Send logs to NervesHub.
+  """
+  @behaviour :logger_handler
+
+  alias NervesHubLink.Extensions.Logging
+
+  # Callback for :logger handlers
+  @doc false
+  @impl :logger_handler
+  def log(%{msg: {:string, unicode_chardata}} = log_event, _) do
+    Logging.send_log_line(log_event.level, unicode_chardata, log_event.meta)
+  end
+
+  def log(_, _) do
+    # Ignore log events which aren't string log lines
+  end
+end